+++ R/datasummary_functions.R
@@ -15,7 +15,7 @@
 Mean <- function(x, fmt = NULL, na.rm = TRUE, ...) {
   out <- mean(x, na.rm = na.rm)
   if (!is.null(fmt)) {
-    out <- rounding(out, fmt)
+    out <- rounding(out, fmt, ...)
   }
   return(out)
 }
@@ -43,7 +43,7 @@ N <- function(x) sprintf("%.0f", length(x))
 Median <- function(x, fmt = NULL, na.rm = TRUE, ...) {
   out <- stats::median(x, na.rm = na.rm)
   if (!is.null(fmt)) {
-    out <- rounding(out, fmt)
+    out <- rounding(out, fmt, ...)
   }
   return(out)
 }
@@ -63,7 +63,7 @@ Min <- function(x, fmt = NULL, na.rm = TRUE, ...) {
     out <- sprintf("%.0f", out)
   }
   if (!is.null(fmt)) {
-    out <- rounding(out, fmt)
+    out <- rounding(out, fmt, ...)
   }
   return(out)
 }
@@ -84,7 +84,7 @@ Max <- function(x, fmt = NULL, na.rm = TRUE, ...) {
     out <- sprintf("%.0f", out)
   }
   if (!is.null(fmt)) {
-    out <- rounding(out, fmt)
+    out <- rounding(out, fmt, ...)
   }
   return(out)
 }
@@ -101,7 +101,7 @@ Max <- function(x, fmt = NULL, na.rm = TRUE, ...) {
 SD <- function(x, fmt = NULL, na.rm = TRUE, ...) {
   out <- stats::sd(x, na.rm = na.rm)
   if (!is.null(fmt)) {
-    out <- rounding(out, fmt)
+    out <- rounding(out, fmt, ...)
   }
   return(out)
 }
@@ -118,7 +118,7 @@ SD <- function(x, fmt = NULL, na.rm = TRUE, ...) {
 Var <- function(x, fmt = NULL, na.rm = TRUE, ...) {
   out <- stats::var(x, na.rm = na.rm)
   if (!is.null(fmt)) {
-    out <- rounding(out, fmt)
+    out <- rounding(out, fmt, ...)
   }
   return(out)
 }
@@ -167,7 +167,7 @@ NUnique <- function(x, ...) sprintf("%.0f", length(unique(x)))
 P0 <- function(x, fmt = NULL, na.rm = TRUE, ...) {
   out <- stats::quantile(x, prob = 0, na.rm = na.rm)
   if (!is.null(fmt)) {
-    out <- rounding(out, fmt)
+    out <- rounding(out, fmt, ...)
   }
   return(out)
 }
@@ -184,7 +184,7 @@ P0 <- function(x, fmt = NULL, na.rm = TRUE, ...) {
 P25 <- function(x, fmt = NULL, na.rm = TRUE, ...) {
   out <- stats::quantile(x, prob = 0.25, na.rm = na.rm)
   if (!is.null(fmt)) {
-    out <- rounding(out, fmt)
+    out <- rounding(out, fmt, ...)
   }
   return(out)
 }
@@ -201,7 +201,7 @@ P25 <- function(x, fmt = NULL, na.rm = TRUE, ...) {
 P50 <- function(x, fmt = NULL, na.rm = TRUE, ...) {
   out <- stats::quantile(x, prob = 0.50, na.rm = na.rm)
   if (!is.null(fmt)) {
-    out <- rounding(out, fmt)
+    out <- rounding(out, fmt, ...)
   }
   return(out)
 }
@@ -218,7 +218,7 @@ P50 <- function(x, fmt = NULL, na.rm = TRUE, ...) {
 P75 <- function(x, fmt = NULL, na.rm = TRUE, ...) {
   out <- stats::quantile(x, prob = 0.75, na.rm = na.rm)
   if (!is.null(fmt)) {
-    out <- rounding(out, fmt)
+    out <- rounding(out, fmt, ...)
   }
   return(out)
 }
@@ -235,7 +235,7 @@ P75 <- function(x, fmt = NULL, na.rm = TRUE, ...) {
 P100 <- function(x, fmt = NULL, na.rm = TRUE, ...) {
   out <- stats::quantile(x, prob = 1, na.rm = na.rm)
   if (!is.null(fmt)) {
-    out <- rounding(out, fmt)
+    out <- rounding(out, fmt, ...)
   }
   return(out)
 }